requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
Pillow==10.1.0
stripe==7.8.0
python-decouple==3.8
drf-yasg==1.21.7
django-filter==23.5
redis==5.0.1
django-redis==5.4.0 